Redesign public site: card-only home + Армія+ dark theme
- Home is now a clean card landing only (header + modality chips + two
  category cards + phone + address). The services catalog and the booking
  form are removed from the first screen.
- New /booking page holds the full flow (services, preparation, booking
  form). The category cards deep-link to it (?category=military|civilian,
  which preselects the patient category).
- Apply an "Армія+"-inspired dark military theme, scoped to public pages
  (.publicShell) so the staff cabinet stays on its light theme: deep olive
  background, olive/khaki surfaces, khaki accent for the civilian card,
  olive-green highlights and dark form fields.
- Update rendered-html tests: home routes into /booking and /staff/login;
  the 38-service catalog now renders on /booking.
