+++ tests/rendered-html.test.mjs
@@ -4,13 +4,13 @@ import test from "node:test";
 const developmentPreviewMeta =
   /<meta(?=[^>]*\bname=["']codex-preview["'])(?=[^>]*\bcontent=["']development["'])[^>]*>/i;
 
-async function renderHome() {
+async function renderPath(path) {
   const workerUrl = new URL("../dist/server/index.js", import.meta.url);
   workerUrl.searchParams.set("test", `${process.pid}-${Date.now()}`);
   const { default: worker } = await import(workerUrl.href);
 
   return worker.fetch(
-    new Request("http://localhost/", {
+    new Request(`http://localhost${path}`, {
       headers: { accept: "text/html" },
     }),
     {
@@ -25,6 +25,8 @@ async function renderHome() {
   );
 }
 
+const renderHome = () => renderPath("/");
+
 test("renders development preview metadata", async () => {
   const response = await renderHome();
 
@@ -36,11 +38,19 @@ test("renders development preview metadata", async () => {
   assert.match(await response.text(), developmentPreviewMeta);
 });
 
-test("renders the full service catalog and same-system staff link", async () => {
+test("home is a card landing that routes into booking and staff login", async () => {
   const response = await renderHome();
+  const html = await response.text();
+  assert.match(html, /Чернігівський військовий госпіталь/);
+  assert.match(html, /href=["']\/booking\?category=military["']/);
+  assert.match(html, /href=["']\/booking\?category=civilian["']/);
+  assert.match(html, /href=["']\/staff\/login["']/);
+  assert.doesNotMatch(html, /radiologyos-app\.adamenko-artem96\.chatgpt\.site/);
+});
+
+test("booking page renders the full service catalog", async () => {
+  const response = await renderPath("/booking");
   const html = await response.text();
   const visibleHtml = html.replace(/<!--.*?-->/g, "");
   assert.match(visibleHtml, /Повний каталог:\s*38\s*послуг/);
-  assert.match(html, /href=["']\/staff["']/);
-  assert.doesNotMatch(html, /radiologyos-app\.adamenko-artem96\.chatgpt\.site/);
 });
